requirements.txt
scrilangchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.22
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
